backend/requirements.txt
# Core numerical & image processing
numpy==1.26.4
opencv-python==4.11.0.86
pillow

# Deep learning
torch
ultralytics

# Segment Anything (SAM)
git+https://github.com/facebookresearch/segment-anything.git

# API server
fastapi
uvicorn
python-multipart
